+++ ExpandWorldData/ExpandWorldData.cs
@@ -11,7 +11,7 @@ public class EWD : BaseUnityPlugin
 {
   public const string GUID = "expand_world_data";
   public const string NAME = "Expand World Data";
-  public const string VERSION = "1.70";
+  public const string VERSION = "1.70.1";
 #nullable disable
   public static EWD Instance;
   public static Harmony Harmony;
@@ -37,7 +37,6 @@ public class EWD : BaseUnityPlugin
     Instance = this;
     Log.Init(Logger);
     Yaml.Init();
-    BiomeManager.SetupBiomeArrays();
     ConfigWrapper wrapper = new("expand_config", Config, ConfigSync, InvokeRegenerate);
     Configuration.Init(wrapper);
     Harmony = new(GUID);
@@ -75,6 +74,7 @@ public class EWD : BaseUnityPlugin
   public void LateUpdate()
   {
     WaterColor.Transition(Time.deltaTime);
+    VegetationLoading.UpdateReload(Time.unscaledDeltaTime);
   }
 
 #pragma warning disable IDE0051
@@ -86,4 +86,3 @@ public class EWD : BaseUnityPlugin
 
 
 }
-